+++ CHANGELOG.md
@@ -19,7 +19,7 @@ Status: Available for use
 - Bump serde_yaml from 0.7.5 to 0.8.4 - PATCH
 - Bump structopt from 0.2.18 to 0.3.13 - PATCH
 - Bump uuid from 0.6.5 to 0.8.1 - PATCH
-- *TEMPORARY* Remove logging verbosity switch to restore working build
+- Rework logging and verbosity setting to use simplelog (needed because of `structopt` upgrade) - PATCH
 
 ### Added
 - Allow the docker-in-docker image to be specified in configuration - MINOR